requirements.txt
boto3>=1.28.0
anthropic>=0.18.0
google-genai>=0.1.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
jinja2>=3.1.0
tabulate>=0.9.0
pyarrow>=14.0.0
